Move profiling ffi crate back out of profiling
Due to a [rust bug][1], LTO was not being applied, and we need the lib
crate type in order to run tests and clippy lints. For now, keep these
as separate crates.

  [1]: https://github.com/rust-lang/rust/issues/51009
