+++ Cargo.toml
@@ -4,6 +4,7 @@
 [workspace]
 members = [
     "profiling",
+    "profiling-ffi",
     "ddcommon",
     "ddcommon-ffi",
     "ddtelemetry",
