+++ LICENSE-3rdparty.yml
@@ -1,5 +1,5 @@
 ---
-root_name: "datadog-profiling, ddcommon, ddcommon-ffi, ddtelemetry"
+root_name: "datadog-profiling, ddcommon, datadog-profiling-ffi, ddcommon-ffi, ddtelemetry"
 third_party_libraries:
   - package_name: aho-corasick
     package_version: 0.7.18
