+++ profiling-ffi/Cargo.toml
@@ -1,0 +1,23 @@
+# Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+# This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
+
+[package]
+name = "datadog-profiling-ffi"
+version = "0.8.0-alpha.1"
+edition = "2021"
+license = "Apache-2.0"
+
+[lib]
+# LTO is ignored if "lib" is added as crate type
+# cf. https://github.com/rust-lang/rust/issues/51009
+crate-type = ["staticlib", "cdylib"]
+
+[dependencies]
+anyhow = "1.0"
+chrono = "0.4"
+datadog-profiling = { path = "../profiling"}
+hyper = {version = "0.14", default-features = false}
+ddcommon = { path = "../ddcommon"}
+ddcommon-ffi = { path = "../ddcommon-ffi"}
+libc = "0.2"
+tokio-util = "0.7.1"
