+++ profiling-ffi/datadog_profiling-static.pc.in
