+++ profiling-ffi/datadog_profiling_with_rpath.pc.in
